% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/packagedocumentationfile.R
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{rstattoys: A package for computating the notorious bar statistic.}
\description{
The rstattoys package is how I'm learning about package development
}
\section{Foo functions}{

The foo functions ...
}

