% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanPerFactor.R
\docType{data}
\name{mean_per_factor}
\alias{mean_per_factor}
\title{Calculate the mean value per factor in a data.frame}
\format{\preformatted{ num 2
}}
\usage{
mean_per_factor
}
\arguments{
\item{data}{A data.frame.}

\item{fact}{A string.}

\item{param}{A number.}
}
\value{
The mean of \code{param} split by the factor, \code{fact}, in \code{data}.
}
\description{
Calculate the mean value per factor in a data.frame
}
\section{Warning}{

This function was written to learn about package namespaces.
}
\examples{
mean_per_factor(data.df, "species")
add(10, 1)
}
\keyword{datasets}

